requirements.txt
feedparser
requests
python-dotenv
rapidfuzz
anthropic
google-genai
pydantic